simplify models and add uri to Disallow
