+++ app/src/main/kotlin/emu/dev/spotify_swipe/api/data/Artist.kt
@@ -1,5 +1,14 @@
 package emu.dev.spotify_swipe.api.data
 
+data class ArtistSimple(
+    val external_urls: Map<String, String> = mapOf(),
+    val href: String?,
+    val id: String?,
+    val name: String?,
+    val type: String?,
+    val uri: String?
+)
+
 data class Artist(
     val external_urls: Map<String, String> = mapOf(),
     val followers: Follower?,
