+++ app/src/main/kotlin/emu/dev/spotify_swipe/api/data/Disallow.kt
@@ -1,14 +1,14 @@
 package emu.dev.spotify_swipe.api.data
 
-enum class Disallow {
-    interrupting_playback,
-    pausing,
-    resuming,
-    seeking,
-    skipping_next,
-    skipping_prev,
-    toggling_repeat_context,
-    toggling_shuffle,
-    toggling_repeat_track,
-    transferring_playback
+enum class Disallow(val uri: String) {
+    INTERRUPTING_PLAYBACK("interrupting-playback"),
+    PAUSING("pausing"),
+    RESUMING("resuming"),
+    SEEKING("seeking"),
+    SKIPPING_NEXT("skipping-next"),
+    SKIPPING_PREV("skipping-prev"),
+    TOGGLING_REPEAT_CONTEXT("toggling-repeat-context"),
+    TOGGLING_SHUFFLE("toggling-shuffle"),
+    TOGGLING_REPEAT_TRACK("toggling-repeat-track"),
+    TRANSFERRING_PLAYBACK("transferring-playback")
 }
