+++ app/src/main/kotlin/emu/dev/spotify_swipe/api/data/Error.kt
@@ -4,3 +4,9 @@ data class Error(
     val status: Int?,
     val message: String?
 )
+
+data class ErrorPlayer(
+    val status: Int?,
+    val message: String?,
+    val reason: String
+)
