--- app/src/main/kotlin/emu/dev/spotify_swipe/api/data/ErrorPlayer.kt
@@ -1,7 +1,0 @@
-package emu.dev.spotify_swipe.api.data
-
-data class ErrorPlayer(
-    val status: Int?,
-    val message: String?,
-    val reason: String
-)
