+++ app/src/main/kotlin/emu/dev/spotify_swipe/api/data/Track.kt
@@ -1,5 +1,25 @@
 package emu.dev.spotify_swipe.api.data
 
+data class TrackSimple(
+    val artists: List<ArtistSimple>?,
+    val available_markets: List<String>?,
+    val disc_number: Int?,
+    val duration_ms: Int?,
+    val explicit: Boolean?,
+    val external_urls: Map<String, String> = mapOf(),
+    val href: String?,
+    val id: String?,
+    val is_playable: Boolean?,
+    val linked_from: TrackLink?,
+    val restrictions: Restriction?,
+    val name: String?,
+    val preview_url: String?,
+    val track_number: Int?,
+    val type: String?,
+    val uri: String?,
+    val is_local: Boolean?
+)
+
 data class Track (
     val album: AlbumSimple?,
     val artists: List<ArtistSimple>?,
@@ -21,4 +41,12 @@ data class Track (
     val type: String?,
     val uri: String?,
     val is_local: Boolean?
+)
+
+data class TrackLink(
+    val external_urls: Map<String, String> = mapOf(),
+    val href: String?,
+    val id: String?,
+    val type: String?,
+    val uri: String?
 )
