--- app/src/main/kotlin/emu/dev/spotify_swipe/api/data/TrackLink.kt
@@ -1,9 +1,0 @@
-package emu.dev.spotify_swipe.api.data
-
-data class TrackLink(
-    val external_urls: Map<String, String> = mapOf(),
-    val href: String?,
-    val id: String?,
-    val type: String?,
-    val uri: String?
-)
